refactor: cleanup project structure
- Cleanup project structure using bulletproof-react as a guideline.
