+++ src/components/NavBar.js
@@ -7,7 +7,7 @@ import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
 import { makeStyles } from '@material-ui/core/styles';
 
-import logoImg from 'images/logo.png';
+import logoImg from 'assets/images/logo.png';
 
 const useStyles = makeStyles(theme => ({
   logo: {
